requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==0.4.3
altair==4.2.0
analytics-python==1.4.0
antlr4-python3-runtime==4.8
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
backoff==1.10.0
bcrypt==4.0.1
blinker==1.5
boto3==1.26.4
botocore==1.29.4
braceexpand==0.1.7
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
datasets==2.4.0
decorator==5.1.1
diffusers==0.3.0
dill==0.3.5.1
dnspython==2.2.1
einops==0.3.0
entrypoints==0.4
fastapi==0.86.0
ffmpy==0.3.0
filelock==3.8.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.10.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
gradio==3.1.4
grpcio==1.50.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==5.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.0
kiwisolver==1.4.4
kornia==0.6.0
-e git+https://github.com/justinpinkney/stable-diffusion.git@fc3a4fd8d5e171627ef30769146c68b9c072dffb#egg=latent_diffusion
linkify-it-py==1.0.3
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.1
mdurl==0.1.2
monotonic==1.6
multidict==6.0.2
multiprocess==0.70.13
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.5.5.64
orjson==3.8.1
packaging==21.3
pandas==1.5.1
paramiko==2.12.0
parquet-cli==1.3
Pillow==9.3.0
postmarker==1.0
protobuf==3.20.3
pudb==2019.2
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydeck==0.8.0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.2
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.4.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywatchman==1.4.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.20.4
streamlit==1.14.0
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
test-tube==0.7.5
tifffile==2022.10.10
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
transformers==4.22.2
twilio==7.15.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uc-micro-py==1.0.1
Unidecode==1.3.6
urllib3==1.26.12
urwid==2.1.2
uvicorn==0.19.0
validators==0.20.0
wcwidth==0.2.5
webdataset==0.2.5
Werkzeug==2.2.2
xxhash==3.1.0
yarl==1.8.1
zipp==3.10.0
